+++ src/test/java/lotto/domain/LottoManagerTest.java
@@ -1,8 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoManager;
+import lotto.domain.LottoSupplier;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
