+++ src/test/java/lotto/domain/LottoStatsTest.java
@@ -1,7 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.LottoStats;
+import lotto.domain.Money;
+import lotto.domain.Rating;
+import lotto.domain.RatingInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
