+++ src/test/java/lotto/domain/RatingInfoTest.java
@@ -1,7 +1,9 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.Rating;
+import lotto.domain.RatingInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
